+++ Runtime/Addressable/AddressableAddressAttribute.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 namespace SaintsField.Addressable
 {
-#if SAINTSFIELD_ADDRESSABLE
+#if SAINTSFIELD_ADDRESSABLE && !SAINTSFIELD_ADDRESSABLE_DISABLED
     public class AddressableAddressAttribute: PropertyAttribute, ISaintsAttribute
     {
         public SaintsAttributeType AttributeType => SaintsAttributeType.Field;