+++ Runtime/Addressable/AddressableLabelAttribute.cs
@@ -2,7 +2,7 @@
 
 namespace SaintsField.Addressable
 {
-#if SAINTSFIELD_ADDRESSABLE
+#if SAINTSFIELD_ADDRESSABLE && !SAINTSFIELD_ADDRESSABLE_DISABLED
     public class AddressableLabelAttribute: PropertyAttribute, ISaintsAttribute
     {
         public SaintsAttributeType AttributeType => SaintsAttributeType.Field;