+++ Runtime/Addressable/AddressableNotEnabled.cs
@@ -1,0 +1,8 @@
+﻿namespace SaintsField.Addressable
+{
+#if !SAINTSFIELD_ADDRESSABLE || SAINTSFIELD_ADDRESSABLE_DISABLED
+    public class AddressableNotEnabled
+    {
+    }
+#endif
+}