+++ Runtime/AiNavigation/AiNavigationNotEnabled.cs
@@ -1,0 +1,8 @@
+﻿namespace SaintsField.AiNavigation
+{
+#if !SAINTSFIELD_AI_NAVIGATION || SAINTSFIELD_AI_NAVIGATION_DISABLED
+    public static class AiNavigationNotEnabled
+    {
+    }
+#endif
+}